Removed junk
The default.blade.php file used to be the homepage, and I just changed some lines to make it into the default layout.
However, this meant that there was some customizations left over that made it act like the homepage, like having the index.html's CSS built-in.

I removed the customizations, as well as added some commentation and directions for how to use the layout.